pyproject.toml
[project]
name = "perplexity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "tavily>=1.1.0",
    "tavily-python>=0.5.0",
    "typing-extensions>=4.12.0",
]
